Move most of PreWarm namespace back to http
